+++ src/main/java/cn/wthee/pcrtool/adapter/EquipmentListAdapter.kt
@@ -12,8 +12,8 @@ import cn.wthee.pcrtool.MyApplication
 import cn.wthee.pcrtool.R
 import cn.wthee.pcrtool.data.db.view.EquipmentMaxData
 import cn.wthee.pcrtool.databinding.ItemCommonBinding
-import cn.wthee.pcrtool.ui.detail.equipment.EquipmentDetailsDialogFragment
-import cn.wthee.pcrtool.ui.home.EquipmentListFragment
+import cn.wthee.pcrtool.ui.tool.equip.EquipmentDetailsDialogFragment
+import cn.wthee.pcrtool.ui.tool.equip.EquipmentListFragment
 import cn.wthee.pcrtool.utils.Constants.EQUIPMENT_URL
 import cn.wthee.pcrtool.utils.Constants.WEBP
 import cn.wthee.pcrtool.utils.ResourcesUtil
