--- src/main/java/cn/wthee/pcrtool/adapter/viewpager/MainPagerAdapter.kt
@@ -1,31 +1,0 @@
-package cn.wthee.pcrtool.adapter.viewpager
-
-import android.util.SparseArray
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
-import androidx.viewpager2.adapter.FragmentStateAdapter
-import cn.wthee.pcrtool.ui.home.CharacterListFragment
-import cn.wthee.pcrtool.ui.home.EquipmentListFragment
-
-class MainPagerAdapter(activity: FragmentActivity) : FragmentStateAdapter(activity) {
-
-    private val mFragments: SparseArray<Fragment> = SparseArray()
-
-    init {
-        mFragments.put(PAGE_CHARACTER, CharacterListFragment())
-        mFragments.put(PAGE_EQUIP, EquipmentListFragment())
-    }
-
-    override fun createFragment(position: Int): Fragment {
-        return mFragments[position]
-    }
-
-    override fun getItemCount(): Int {
-        return mFragments.size()
-    }
-
-    companion object {
-        const val PAGE_CHARACTER = 0
-        const val PAGE_EQUIP = 1
-    }
-}
