+++ src/main/java/cn/wthee/pcrtool/ui/character/attr/CharacterAttrViewModel.kt
@@ -1,4 +1,4 @@
-package cn.wthee.pcrtool.ui.detail.character.attr
+package cn.wthee.pcrtool.ui.character.attr
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
