+++ src/main/java/cn/wthee/pcrtool/ui/character/attr/CharacterAttrViewModelFactory.kt
@@ -1,16 +1,21 @@
-package cn.wthee.pcrtool.ui.detail.equipment
+package cn.wthee.pcrtool.ui.character.attr
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import cn.wthee.pcrtool.data.db.repository.CharacterRepository
 import cn.wthee.pcrtool.data.db.repository.EquipmentRepository
 
-class EquipmentDetailsViewModelFactory(
+class CharacterAttrViewModelFactory(
+    private val repository: CharacterRepository,
     private val equipmentRepository: EquipmentRepository
 ) : ViewModelProvider.NewInstanceFactory() {
 
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return EquipmentDetailsViewModel(equipmentRepository) as T
+        return CharacterAttrViewModel(
+            repository,
+            equipmentRepository
+        ) as T
     }
 }
