+++ src/main/java/cn/wthee/pcrtool/ui/character/skill/CharacterSkillFragment.kt
@@ -1,4 +1,4 @@
-package cn.wthee.pcrtool.ui.detail.character.skill
+package cn.wthee.pcrtool.ui.character.skill
 
 import android.os.Bundle
 import android.view.LayoutInflater
