+++ src/main/java/cn/wthee/pcrtool/ui/tool/equip/EquipmentDetailsDialogFragment.kt
@@ -1,4 +1,4 @@
-package cn.wthee.pcrtool.ui.detail.equipment
+package cn.wthee.pcrtool.ui.tool.equip
 
 import android.annotation.SuppressLint
 import android.content.DialogInterface
