+++ src/main/res/drawable/card_background.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <corners android:radius="@dimen/card_radius" />
+    <solid android:color="@color/colorWhite" />
+
+</shape>
