+++ src/main/res/layout/activity_main.xml
@@ -4,7 +4,6 @@
     android:id="@+id/main_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:backgroundTint="@color/colorGray"
     android:animateLayoutChanges="true">
 
     <androidx.fragment.app.FragmentContainerView
