--- src/main/res/layout/fragment_main_pager.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.motion.widget.MotionLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:animateLayoutChanges="true"
-    android:orientation="vertical"
-    android:background="@color/colorBg"
-    app:layoutDescription="@xml/fragment_main_pager_scene">
-
-
-    <include
-        android:id="@+id/main_toolbar"
-        layout="@layout/view_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/layout_tab"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/tab_height_1"
-        app:layout_collapseMode="pin"
-        app:layout_constraintTop_toBottomOf="@id/main_toolbar"
-        app:tabInlineLabel="true"
-        app:tabMode="fixed" />
-
-    <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/no_data_tip"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginTop="@dimen/s_margin"
-        android:padding="0dp"
-        android:text="@string/data_null_character"
-        android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle1"
-        android:textColor="@color/colorPrimary"
-        android:visibility="gone"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/layout_tab"
-        tools:visibility="visible" />
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/main_view_pager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constrainedHeight="true"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/no_data_tip" />
-
-</androidx.constraintlayout.motion.widget.MotionLayout>
