+++ src/main/res/layout/fragment_tool_news.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
     android:background="@drawable/background_color"
     android:orientation="vertical">
 
@@ -10,23 +10,23 @@
         android:id="@+id/tool_news"
         layout="@layout/view_toolbar" />
 
-    <com.google.android.material.card.MaterialCardView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+    <com.google.android.material.tabs.TabLayout
+        android:id="@+id/tabs"
         android:layout_marginStart="@dimen/xm_margin"
         android:layout_marginEnd="@dimen/xm_margin"
-        app:cardCornerRadius="@dimen/card_radius">
-
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/tabs"
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/tab_height_1" />
-
-    </com.google.android.material.card.MaterialCardView>
+        android:background="@drawable/card_background"
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/tab_height_1" />
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/view_pager"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent"
+        tools:visibility="gone" />
+
+    <include
+        layout="@layout/fragment_tool_news_list"
+        android:visibility="gone"
+        tools:visibility="visible" />
 
 </LinearLayout>
