+++ src/main/res/layout/item_news.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/news_item"
@@ -8,42 +8,37 @@
     android:layout_marginStart="@dimen/m_margin"
     android:layout_marginEnd="@dimen/m_margin"
     android:layout_marginBottom="@dimen/m_margin"
-    app:cardCornerRadius="@dimen/card_radius">
+    android:background="@drawable/card_background"
+    android:foreground="?selectableItemBackgroundBorderless"
+    android:minHeight="@dimen/news_card_height"
+    android:orientation="vertical">
 
-    <LinearLayout
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/tags"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:minHeight="@dimen/news_card_height"
-        android:orientation="vertical">
+        android:layout_height="wrap_content"
+        android:nestedScrollingEnabled="false"
+        android:orientation="horizontal"
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+        tools:itemCount="3"
+        tools:listitem="@layout/item_tag" />
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/tags"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:nestedScrollingEnabled="false"
-            android:orientation="horizontal"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            tools:itemCount="3"
-            tools:listitem="@layout/item_tag" />
+    <com.google.android.material.textview.MaterialTextView
+        android:id="@+id/news_title"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:padding="@dimen/xm_padding"
+        android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle1"
+        android:textColor="@color/colorPrimary"
+        tools:text="@string/app_name" />
 
-        <com.google.android.material.textview.MaterialTextView
-            android:id="@+id/news_title"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:padding="@dimen/xm_padding"
-            android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle1"
-            android:textColor="@color/colorPrimary"
-            tools:text="@string/app_name" />
+    <com.google.android.material.textview.MaterialTextView
+        android:id="@+id/news_date"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end"
+        android:padding="@dimen/xm_padding"
+        android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"
+        tools:text="@string/app_name" />
 
-        <com.google.android.material.textview.MaterialTextView
-            android:id="@+id/news_date"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:padding="@dimen/xm_padding"
-            android:layout_gravity="end"
-            android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"
-            tools:text="@string/app_name" />
-
-    </LinearLayout>
-
-</com.google.android.material.card.MaterialCardView>
+</LinearLayout>
