+++ src/main/res/layout/item_skill.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_marginBottom="@dimen/m_margin"
-    android:elevation="@dimen/elevation_card">
+    android:elevation="@dimen/card_elevation">
 
     <com.google.android.material.textview.MaterialTextView
         android:id="@+id/name"
