+++ src/main/res/values/styles.xml
@@ -36,4 +36,11 @@
         <item name="android:background">@drawable/dialog_background</item>
     </style>
 
+    <!-- 圆角图片 -->
+    <style name="roundedCornerStyle">
+        <item name="cornerFamilyTopLeft">rounded</item>
+        <item name="cornerFamilyTopRight">rounded</item>
+        <item name="cornerSize">@dimen/card_radius</item>
+    </style>
+
 </resources>
